Many updated scripts and services udev runs fast_udev script, mountscript too slow for udev fast_udev starts udev service mount.conf to pass vars to udev service udev services runs mountscript eve.py and eve_clean.py follows wan interface interface blink service and script

+++ mountscript.sh
@@ -1,26 +1,39 @@
 #!/bin/bash
 # script for Eve to mount and read for updated bpf programs
-#
-# format is /dev/sdc sdc 1 /media/sdc1DEVICENAME, to work with udev rule
-# rule is something like
-# KERNEL=="sd[abcde]", SUBSYSTEM=="block", SUBSYSTEMS=="usb", ATTRS{product}=="Cruzer Fit", ATTRS{serial}=="", \
-#       ACTION=="add", RUN+="thisscript $devnode $name 1 $env{ID_MODEL}"
 
-systemd-mount --no-block --automount=yes --collect $1$3 /media/$2$3$4
+LOG="/eve/mount_log"
+set +e
 
+echo "mounting" >> $LOG
+if [ ! -d /media/$2$3$4 ]; then
+    mkdir /media/$2$3$4
+fi
+mount $1$3 /media/$2$3$4
+#systemd-mount --no-block --automount=yes --collect $1$3 /media/$2$3$4
+# give it a second to mount
+sleep 1
+
+echo "declaring vars" >> $LOG
 MOUNT="/media/$2$3$4"
 EVE_PY="eve.py"
 EVE_C="eve.c"
 CLEAN="eve_clean.py"
 
-if [ -f "$MOUNT/$EVE_PY" ]; then
+echo "checking for eve.py and eve.c" >> $LOG
+if [ -f "$MOUNT/$EVE_PY" -a -f "$MOUNT/$EVE_C" ]; then
+    echo "found eve.py and eve.c! copying" >> $LOG
     cp $MOUNT/$EVE_PY /eve/$EVE_PY
     cp $MOUNT/$EVE_C /eve/$EVE_C
-    systemctl restart eve.service
-    ./eve/blinkLED.sh 1
+    echo "restarting eve service..." >> $LOG
+    systemctl stop eve.service
+    systemctl start eve.service
+    bash /eve/blinkLED.sh 1
 fi
 
 # also check for the clean program, which is called by eve.service ExecStop
+echo "checking for $MOUNT/$CLEAN" >> $LOG
 if [ -f "$MOUNT/$CLEAN" ]; then
+    echo "eve_clean found! copying" >> $LOG
     cp $MOUNT/$CLEAN /eve/$CLEAN
+    bash /eve/blinkLED.sh 1
 fi
